Clean up CLI code
